create character_mutations and change scope
