+++ src/melee.cpp
@@ -1812,7 +1812,7 @@ static damage_instance hardcoded_mutation_attack( const player &u, const trait_i
     return damage_instance();
 }
 
-std::vector<special_attack> player::mutation_attacks( Creature &t ) const
+std::vector<special_attack> character_mutations::mutation_attacks( Creature &t ) const
 {
     std::vector<special_attack> ret;
